+++ src/utils.c
@@ -21,7 +21,7 @@ t_point create_t_point(double x, double y)
 	return (p);
 }
 
-t_range create_t_range(double start, double end)
+t_range create_t_range(int start, int end)
 {
 	t_range range;
